requirements.txt
matplotlib==3.2.2
torch==1.7.0
numpy==1.21.2
mxnet==1.9.0
opencv_python==4.5.5.62
torchvision==0.8.0
tqdm==4.62.3
Pillow==9.0.1
pydensecrf==1.0rc3
scikit_learn==1.0.2
skimage==0.0
